// src/utils/enums.tsx
export enum AvailableFilters {
  Sex = 'sex',
  Query = 'query',
  Centuries = 'centuries',
  Sort = 'sort',
  Order = 'order',
}

export enum SortOptions {
  Sex = 'sex',
  Born = 'born',
  Died = 'died',
  Name = 'name',
}
export enum FilterTypes {
  All = '',
  Male = 'm',
  Female = 'f',
}

export enum OrderType {
  desc = 'desc',
  asc = 'asc',
}